Move collections to t heir own files.
